refactor: Update dependencies and remove unused imports
This commit updates the project dependencies and removes unused imports. It also refactors the code to remove the injection of `SqlDriver` and replace it with appropriate factories. This change ensures better maintainability and separation of concerns.

Interface changes:
- `CombustivelRepositoryFactory` is introduced, implementing `ICombustivelRepositoryFactory` interface.
- `VeiculoRepositoryFactory` is introduced, implementing `IVeiculoRepositoryFactory` interface.
- `ICombustivelViewModelFactory` is introduced, replacing `CombustivelViewModelFactory` interface.
- `IVeiculosViewModelFactory` is introduced, replacing `VeiculoViewModelFactory` interface.
- `ICombustivelFormViewModelFactory` is introduced, replacing `CombustivelFormViewModelFactory` interface.
- `IVeiculoFormViewModelFactory` is introduced, replacing `VeiculoFormViewModelFactory` interface.

The changes improve code organization and make it easier to manage the dependencies within the project.